Create local library app
The app is being created using node and express as the main framework.
This website leverages the database from MongoDB Atlas to store a
collection of books, book instances, authors and genres, which
are related by means of specific fields.

At the moment, the different routes for the project have been defined
along with the controllers, where each section is located in separate
modules.

On the other hand, the the template views for the list of each
instance of models have been created with the help of ejs. Likewise,
only the controllers to fetch all the instances in each model have been
defined.